creating a place for utility functions that are shared between the client and the server
