+++ public/modules/wagers/controllers/wagers.client.controller.js
@@ -40,42 +40,8 @@ angular.module('wagers').controller('WagersController', ['$scope', '$stateParams
 			});
 		};
 
-		function calcWinnings (amount, juice) {
-			juice = juice.toString();
-			juice = juice.trim();
-			var slashPosition = juice.indexOf('/');
-			var winnings = 0;
-			var conversion = 0;
-
-			if (juice === 'EVEN') {
-				winnings = amount;
-			}
-			else if (slashPosition > 0) {
-				var numerator = juice.split('/')[0];
-				var denominator = juice.split('/')[1];
-				var theOdds = parseFloat(numerator / denominator);
-				console.log(theOdds);
-				winnings = amount * theOdds;
-			}
-			else if (juice[0] === '+') {
-				juice = parseInt(juice);
-				conversion = juice / 100;
-				winnings = conversion * amount;
-			}
-			else {
-				juice = parseInt(juice.substring(1));
-				conversion = 1 / (juice / 100);
-				winnings = conversion * amount;
-			}
-
-			var pay =  amount + winnings;
-			pay = parseFloat(pay).toFixed(2);
-			winnings = parseFloat(winnings).toFixed(2);
-	
-			return {'pay' : pay, 'winnings' : winnings};
-		}
 		$scope.calcWinnings = function(amount, juice) {
-			return calcWinnings(amount, juice).pay;
+			return window.serverBrowserUtils.calcWinnings(amount, juice).pay;
 		};
 
 		$scope.winnings = '';
@@ -83,7 +49,7 @@ angular.module('wagers').controller('WagersController', ['$scope', '$stateParams
 		$scope.onAmountChanged = function() {
 			if (this.amount > 0) {
 				var juice = document.getElementById('itemJuice').innerText;
-				var ret = calcWinnings(this.amount, juice);
+				var ret = window.serverBrowserUtils.calcWinnings(this.amount, juice);
 				$scope.pay = ret.pay;
 				$scope.winnings = ret.winnings;
 			}
